Refactor channel-info-card to assume channel always exists.
ChannelInfoCard component should only be used in places where we can be
reasonably certain that the channel it's trying to display actually
exists (e.g. channel search results, channel mentions, etc.).

Places that use ChannelInfoCard and are not sure if the channel exists
(e.g. landing on the non-existing channel page directly), should handle
the errors of channel not existing outside the channel-info-card.