Add a new (better) version of the channel info card. (#907)
Mostly plagiarized from the league stuff ><

I haven't actually re-used anything from the league components, as that
just seems like a cause for a headache when those designs inevitably
diverge.

However, maybe *some* things could be re-used? The badge stuff in
particular is like 99% the same, and could perhaps be made into a more
general thing. The only problem is I don't know what to call and where
to put that more general thing..

* Don't handle erros in join-channel dialog.

This also fixes an existing issue caused by manually setting an error
text to the input field, which would override any validation errors.

* Refactor channel-info-card to assume channel always exists.

ChannelInfoCard component should only be used in places where we can be
reasonably certain that the channel it's trying to display actually
exists (e.g. channel search results, channel mentions, etc.).

Places that use ChannelInfoCard and are not sure if the channel exists
(e.g. landing on the non-existing channel page directly), should handle
the errors of channel not existing outside the channel-info-card.