Create a channel list page.
Currently there's no way to access this page. But the plan is to have
the join-channel button navigate to this page as soon as the
create-channel button is added to it.

This channel list is meant to be used by both the admins and the public
users. Admins will have additional actions that they can use on a
channel, e.g. open admin view of a channel.