Create a channel list page. (#925)
This channel list is meant to be used by both the admins and the public
users. Admins will have additional actions that they can use on a
channel, e.g. open admin view of a channel.

* Add a dedicated page to create a channel.

This page currently looks pretty bare (basically what we had previously
in the join-channel dialog), but gonna add more stuff to it in the next
PR.

This commit also replaces the chat channels subheader with a clickable
one which navigates the user to the channel list, instead of opening the
join-channel dialog.

Also, a button to the create-channel page is added to the channel-list
page.

* Add the tests for searchChannels method.