Fix a bug in infinite-scroll-list which would sometimes not load chat messages.
The problem was when the initial state of `hasPrevData`/`hasNextData`
was `false` (or `undefined`), which was the case in receiving the
`chatReady` event *before* the chat `init` event. So we would start
rendering the channel and channel messages with the default state and
`hasMoreHistory` would be `undefined`. This meant that the
inifnite-scroll-list would not render the observing element
(i.e. LoadingArea) and nothing would be observed.

Once we've received the `init` event and `hasMoreHistory` would become
`true`, the observing element would finally render, but there was
nothing in the infinite-scroll-list that would try to observe the
element again, and it would get stuck like that. Now the
infinite-scroll-list automatically restars observing if
`hasPrevData`/`hasNextData` changes.