Use `resetDefaultValueLocale` in the i18next-parser.
This option ensures that the default values we use in code are the
source of truth for our translation files, so in case any of the default
values are updated, the translation files gets updated as well. Note
that this only applies to our main locale, other locales will need to be
updated manually. It would be nice if Weblate marks the translations in
other locales as "needs editing" if the default value of the main locale
changes for that key. Maybe it does already?

Already caught a stale translation string with this feature, so hooray!

Also, I've turned on the `failOnWarnings` option as well. I'm not sure
what types of warnings the i18next-parser gives out, but one of them is
if you use the same key with different default values. So it seems safer
to treat that as an error, I think.