Add automatic language detection and a way to change language for signed out users.
The detected language is first detected from the "navigator" settings
and saved into the local storage and is not currently being sent to the
server during the signup process. That will come in the next commit.

User can change the language on all of the pages that render top-links;
basically splash/faq/login/signup.