Upgrade to latest TS + fix issues (since i18next no longer crashes it).
Had to fork tsyringe to get the latest changes for the time being, this
can be undone once https://github.com/microsoft/tsyringe/pull/226 is in
a published version.

Prettier and/or our import sorter also seemingly changed and wanted to
reorder capital letters so that is the majority of the changes here.