Utilize react-email for email templates + use templates via the mailgun API instead of running handlebars locally. (#976)
Mailgun's HTML->Text process seems to use some text-based browser for rendering which looks a bit... weird, but is mostly functional and I imagine is a rare thing for people to see anyway. I wish we could specify a text template alongside the HTML one but unfortunately we cannot. Perhaps in the future we could look at moving to some other service that allows more configuration around these (there's a number of similar services out there, most of them just have more expensive base pricing but scale slightly better).

This PR does not include translation for emails, but the strings are ready for translation (need to setup a new Weblate project for these). The plan is to generate separate templates for each language we support, and upload each of those versions to Mailgun separately. (We need to be careful about the number of templates we use since we have a limit of 100 templates total, so 10 templates * 10 languages would hit that limit, for instance).