Fix the issue where the channels can potentially end up in wrong order.
The previous code could produce a bug in a following case:
- connect with one client to the server and create a channel
- connect with the second client to the server
- second client will receive both the `chatReady` and `init3` events in
  the non-consistent order (depends on how things execute on the server)
- if the second client receives the `init3` event before the `chatReady`
  event, their channel list will be out of order

This fix makes it so we don't send the `init3` event when they connect
to the server, but only immediatelly after they join a channel.

Additionally, we could also change the reducer on the frontend to
always rewrite the channel list when it receives the `chatReady` event,
but it seems unnecessary to me.