Add a dialog to show replay info, use that instead of directly launching dragged-in replays.
This pulls the component used in the local replay browser out to be used
in the dialog as well, with some refactoring to support the style
differences in both cases. I also swapped the sides of the display (it
felt weirdly weighted with the map image on the the right), removed a
bunch of internal padding that shouldn't have been present, and
tightened up the player names.