Allow SB to be registered to open *.rep files.
This is done by 1) adding the currently running program to the list of
program IDs in the Windows registry, and 2) adding that program ID to
the .rep file extension's list of opening program IDs (creating the .rep
extension info if it doesn't exist).

It does not directly associate SB as the default program (or have
anything in our settings for that) because Windows tries to avoid
programs doing that directly now. If no other program has registered for
.rep files, SB will act as the default after registration. If other
programs are registered, Windows should display a dialog to let user
choose the correct program.

SB's program ID will depend on the version that is running (prod,
staging, etc.) so we don't have a lot of different versions fighting for
the file association.

This change also includes a setting to immediately launch replays upon
opening (the default being that we show the same info dialog as when you
drag a replay into the client). It'd be way cooler if SC:R's launch
speed could be sped up to under a second, but is still kind of cool now :)

Closes #769.