Move auth-reducer to immer.
Moves a few of the utilities into the same file as well, and generally
tries to convert state checks into hooks where possible instead of
making them pass their own state in. Highly possible this change will
break *something* (finding problems in non-TS code is harder) but
hopefully I caught everything.

In the future it'd be nice to remove the error stuff in this reducer and
inline the functions that exist in that file, would probably simplify
some things.