Fix replay file association not working if the app wasn't already running.
The `did-finish-load` handler we attached never ran because the promise
awaited in `createWindow` resolved at the same time it would have. Thus,
this also fixes the maximized state not being set properly on load
(since that was meant to run in the same handler).

Additionally, this fixes a few other things about replay loading:
- Replay loading now works while logged out
- Opening a replay will now show the SB window
- Passing relative paths to the ShieldBattery exe will now work
  properly.

Fixes #998.