Add admin panel to upload, edit, and view channel content.
For now, the only channel content that we support is channel banners
(bigger images that channel owners can use to style their channels).
Once we have everything working for the channel banners, the next thing
would be to support channel badges (smaller images that serve as channel
logos) as well.