Add a working integration test for a password change in the new settings flow.
This also adjusts how useAutoFocusRef works by default, as it used to
have a delay for animation reasons but this doesn't really seem to be
necessary in newer browsers/with changes to our animation stuff.
Removing the delay reduces some flakiness in tests and generally makes
things behave better if users interact with things very quickly. I also
adjusted the code to handle cases where we utilize the ref in a
different component than we created it in (the delay before focus won't
apply until the ref is actually set now).