Add support for logging to Datadog.
This adds a custom tracing subscriber that ingests events, collects them
into a 5 second batch, and sends them to Datadog as JSON. Originally I
attempted to use one of the libraries for this, but it had a number of
problems (most notably it only recorded Span's data right when they were
created, so any fields that were updated after construction would not be
set properly when sent to Datadog).

I ended up writing a custom one, but was able to reuse the JsonStorage
stuff from the Bunyan formatter we we're already using, which made it a
lot easier to convert a given span/event to JSON.

The logs here will only be *so* useful for any given request given that
they'll mostly all be to the same /gql path. We may want to think about
ways to get more info about what specifically was being requested
(although any log events from deeper in the app, like in a Query
handler, will include Spans with more info).

Fixes #1012.