Add support for new advanced Lobby settings: turn rate + legacy unit limits.
We already had most of the necessary code on the game side (and
everything reversed), just needed to add new UI in the app and plumb
that through the server -> game loader -> active game manager -> game.

As part of doing that I fixed a number of issues in the create lobby
flow, notably using map IDs in reducers instead of full map state and
greatly simplifying + modernizing the components for selecting a map in
that flow.

I'd like these settings (and the map itself) to be something you can
change while a lobby is open, but we should probably modernize all the
lobby infrastructure before we put in that work, so for now these are
static once the lobby is created.

We may also need to provide more explanation for what these are/what the
values do, but I guess we can see what the response to them is.

Fixes #1025. Fixes #897.