Rewrite game result submission code in the game + add tests.
The app will now submit an empty game result report on game crash (if no
results had been submitted yet). The server will now wait for reports
from everyone before reconciling (instead of just waiting for terminal
results) as in the case of a drop in a 1v1, we might get reports with
conflicting terminal results in succession, and that should trigger a
dispute rather than a victory for whoever was first.

This will mean results might be delayed longer in the case of total
disconnects from one player. The near-future work to ping the webserver
during games (so we can more closely track when games are complete)
should let faster reconciliation happen again in those cases, but for
now we can just accept the slower (but correct) reconciliation.

Also adds a test for the double-loss issue just to make sure we don't
regress on that in the future, although current clients should no longer
send reports that result in that.

Fixes #541. Fixes #1028. Fixes #918. Fixes #849.