Write a replacement for sql-template-strings.
This replacement has better/safer support for adding raw strings to
queries, makes queries immutable, and provides a utility for
concatenating multiple query parts with a separator. I've tried to
update everything that was using the old one to the new patterns, but no
guarantees I hit everything optimally.

Fixes #1044.