Only reuse values if the piece is reused.
Trying to reuse more than this can cause errors, e.g. if we pass
null/undefined for multiple columns and the columns have different
types. If someone is passing a large amount of data, they can choose to
make this optimization manually now by creating a sql fragment for the
data.