Remove code related to the channel content.
Our initial thinking was for server admins to be in charge of uploading
and maintaining channel content that channel owners could use. This
would've eliminated the need to moderate channel content. However,
after investigating some of the services which provide automatic
nsfw-filtering for images, we've decided to simplify this setup and just
allow users to upload their content directly.

This also removes the necessity to treat all channel content as a
separate entity, and can just be saved as an image URL in the channels
settings.

This PR is basically a revert of #999 while keeping some of the general
changes from the PR in. Also, this PR fixes a bug that was introduced in
that PR regarding the chat admin routes.