Add new auth system based around Authorization headers + JWT tokens.
This improves a number of things on the old system:
- Makes it so we no longer need to track user's active sessions to
  remove them (all active sessions for a user are under the
  `sessions:<user_id>:` prefix in redis.
- No longer stores any cached user data in the session key in redis (we
  instead cache this info at `users:<user_id>`, on a shorter timeout
  than the session timeout).
- Tokens can be generated by any server (node or Rust) so we can move
  the login or signup code into Rust, and also refresh tokens from Rust.
- Fixes some race conditions around session destruction if multiple
  requests were in flight for the same session (or if a user had
  requests in flight while they were being banned).

I've removed the redis migration stuff because we weren't really using
it (we generally migrate stuff at runtime there), and the only migration
was related to the old session things that no longer exist.

In the near future I'm going to utilize redis pub/sub to notify changes
to the cached user data, so that our servers can deal with the other
server updating things like email verification properly.