Refresh the current session every ~1 day while the client remains open.
This will prevent the token from randomly expiring if people leave their
clients open for a long time (and there's a lot of leeway here since our
session expirations are pretty long, so a few failures due to not having
connection or something is probably fine). Could make it more resilient
to not having internet connection by using the connection status APIs,
but I think for now this is okay.

Fixes #1048.