Add support for Redis pub/sub, add a GQL mutation for updating permissions + broadcast that change over redis pub/sub.
No support for subscribing to pub/sub on the Rust side at the moment as
I don't have a direct use for it and it'll be more work than doing that
on the JS side, but we can add it in the future if we need it.

Nothing on the client uses this GQL mutation currently (but I will
update the admin page soon), but I've tested it via the GQL playground.

This commit also fixes a usage of redis's KEYS command (moving it to
SCAN) since we've realized that can have some very negative performance
characteristics.

Fixes #1049.