Remove subpaging from the channel settings dialog.
Turns out having subpages in a dialog where some pages might have a Save
button and others don't would work out pretty weird. So the easier
thing is just removing the subpages and make a dedicated dialog for each
subpage (of which we'll only have 2 in the near future most likely).