Fix a bug in channel list search leading to potentially not searching.
This fixes a bug that could've happened if the previous request was
still loading, the new one wouldn't be executed because
infinite-scroll-list only runs the load more data callback if it's not
already loading.

This is a band aid fix; the proper fix would involve fixing
infinite-scroll-list, but is also a bit more work so this will have to
do for now.