Update nginx proxy config to not proxy requests.
I noticed that for map uploads (and image uploads probably) we were getting a warning about buffering requests to a file. Doesn't seem particularly useful/necessary to ever do this (or even buffer them in RAM) since it's all the same machine, so I'm updating the config to just pass directs on directly.

It seemed reasonable to disable max body size on the nginx side as well in this case, which will allow the proxied server to pick whatever limits it wants.