Make launching 64-bit exe work
Just required various minor fixes to code that still assumed 32-bit and
a new sb_init.dll.

Also fixes a bug where in 32-bit exe we'd corrupt net_player_to_game[-1]
and net_player_to_unique[-1] in bw's memory. It happened to be an offset by
u32::MAX and not literal -1, making it crash on 64-bit.