Use Electron's net module to fetch maps instead of got.
This should hopefully fix issues where people are able to access the
APIs and such in-client but then map downloads silently fail in the
background. Fixes #1082.

If this should fail, the handling is still not great and quite silent to
users (unless they submit logs), I'll file a separate bug to add a
dialog or something for this case.