Properly clean up state when users leave during countdown/loading.
So there are 2 issues that were happening in the previous code when user
would leave the lobby:
- We would publish the `cancelLoading` event only to the user who left,
  instead of the whole lobby.
- We would delete the lobby from the loading state only in case everyone
  left, instead of always.

I made the loading cleanup a bit more consistent now and similar to
how/when cancel countdown happens. The only slightly weird thing here is
that we publish the event to the lobby list to add the lobby back to the
list when the loading gets canceled. However this is also done when
cleaning up the countdown state as well, so we have two events fire each
time, because those two cleanup functions are called in unison. We could
perhaps combine these two functions into one, but I don't think it's a
big deal to publish the event twice either?

I also fixed a rare bug that would happen if the last user would leave
the lobby during countdown/loading (this is possible if you start the
game with a computer and leave). The lobby would get deleted, but the
cancel countdown/loading cleanup functions would publish the "add" event
back to the lobby list, making the clients show a non-existent lobby.
This wasn't a huge issue since it would only show the non-existent lobby
if someone else already had the lobby list open. As soon as they close
it and open it again they would get a full lobby list without the
non-existent lobby.