Update aws-sdk to v3. (#1093)
* Update aws-sdk to v3.

Aws sdk v2 has been deprecated:
https://aws.amazon.com/blogs/developer/announcing-end-of-support-for-aws-sdk-for-javascript-v2/

Even though there's still a year until the end-of-support, I thought I'd
be proactive and update to v3 now.

This was mostly done by running their codemod and fixing a couple of
issues manually, e.g.

`npx aws-sdk-js-codemod -t v2-to-v3 ./server/lib/file-upload/aws.ts`

One thing I have *not* done yet is setup DO spaces to be use locally
which would actually test this code. That seems like a pretty important
step tho :d. However, in my defense, setting up that locally is a hassle
and I forgot how to do it, so we can just leave this in open PR now
until that is done.

* Make the new AWS SDK actually work.

Nothing major was broken, but did need a couple of changes:
- `endpoint` requires a "https://" prefix now because the SDK will try
  to create new URL() out of it. This also impacts in how we're
  constructing the public URLs a little bit (need to use the hostname).
- the upload function of the SDK doesn't work with BufferLists anymore
  (which is what our map images were using). So just needed to create a
  Readable stream out of it.

And add a check which might prove useful if we ever move off of DO (you
never know!)