Refactor the admin map manager component to not use the file paths.
`File.path` property is not available anymore in newer versions of
Electron, so we can't use it anymore. However, it shouldn't be needed to
use it in the first place when the user selects files from their disk,
so the map manager component doesn't rely on it anymore.

Also rewrite it in TS/function components.