Update Blizzard protobuf file to latest TS protobuf generation + fix for typechecking.
I do not think the long override is actually *necessary* (from the
previous commit I cherry-picked from the effort to use TSC) as that I
believe was a fix for the GRPC files and they're no longer being built
with TS the same way, but it was a pain to figure out so we can keep it
for now just in case.

Also hold rally-point deps back since the new ones are ESM and we cannot
use them from a non-ESM project.