Add support for 1v1 Fastest matchmaking.
This is similar to 1v1 matchmaking, allowing for alternate race
selection for the random matchup, but introduces a new type of map
selection: positive map selection. In this format, you select which
maps you would like to play on (rather than vetoing) and only opponents
that select the same maps will be valid for matching. This effectively
matches the current map system for the previous community ladder for
Fastest (SCW).