Fix issue with user profiles failing to load if you click into them from the overlay before the overlay loads.
The logic for coalescing with AbortSignals was incorrect here, we need
to avoid aborting any fetch for which there is an un-aborted request
that has been coalesced. I've added a new BatchedAbortSignals class
(effectively an implementation of an `AbortSignal.all` type thing that
can have it's underlying list of signals changed dynamically).

I'm sure this bug exists in a bunch of places in the app, I will follow
a follow-up bug to reproduce this implementation elsewhere.

Fixes #828.