Use redis for ladder rankings instead of a postgres view.
This means rankings are now always computed instantaneously, but also
require looking up User IDs in redis followed by looking up the
rating+user info in the DB. I don't think this should really cause
performance issues but at some point we'll need to paginate this stuff
properly (search queries will probably get weird with that, but I guess
we'll deal with that when we get there :) ).

To migrate to the new system, the server will perform a full update on
startup if the redis keys for the current season are not present. If we
ever need to force a regeneration for whatever reason, we can simply
delete the keys in redis and restart the server.

Old keys don't get cleaned up currently. I don't think this'll really
matter very much as this storage is likely extremely small (and we don't
rotate seasons often), but we can consider doing that in the future.

This fixes #929. Matchmaking rating updates are now performed so long as
the results are reconciled within 6 hours of the season ending. Final
rankings (for historical view) will be generated after this time (I
haven't written code to do that yet, that will be in a follow-up
commit).