Store finalized rankings for seasons in the DB after they have completed.
These are calculated from the DB at the time of finalizing (which at
this point should always happen within 9 hours of the season ending). As
part of this change I adjusted the queries around seasons to include an
end date (which is calculated using the LEAD() function to grab the next
row's start date). This is only slightly useful for this change, but
will be needed for bonus pool changes coming in a follow-up commit so it
seemed like a good time to do it :)

Currently this table only stores the rank, you need to look into
matchmaking_ratings to get the other info. This might be inefficient for
grabbing the whole table? If so, we could put more data in the table, it
just seemed nicer to have them separate for the time being to avoid
needing to migrate this table every time we make changes to
matchmaking_ratings.