Use Points instead of MMR for deciding ranked divisions.
Bonus pool is added to the boundary of each division (so not playing for
a long time can result in division demotions). This will hopefully be
less confusing for people, as the ladder rankings will now also follow
division rankings. It will also hopefully feel like there is more
progression within a season, since you always start at the lowest rank
and progress your way up, regardless of your MMR (although obviously MMR
will affect your peak division).

MMR is still used for matchmaking, so there may be some confusion in the
matches that are chosen (e.g. a bronze player playing a diamond player),
so we may need to be careful about the information we show on pre-match
screens.

We'll need to adjust the division buckets most likely (currently Bronze
1 takes forever to get out of) but we can do that in a follow-up commit.