Fix race condition with setting up subscriptions on a client that immediately disconnects.
One way this could be seen is having a client that immediately crashes
upon viewing the lobby route. In this case, the server would set up all
the subscriptions for this client, but since it had `await`s after
grabbing the client, the client could disconnect before any
subscriptions actually occur. Thus, the cleanup on those subscriptions
that would be assumed to run would also not run.

In this particular example, the client would get locked in a gameplay
activity they couldn't leave, and the lobby (despite actually being
empty) would remain open.

I've made a couple of changes here, for one, clients that are already
disconnected cannot be registered in a gameplay activity (this throws an
error now). This is more of a safety mechanism. The other change fixes
the bug here: when attaching subscriptions to groups that are already
disconnected, the cleanup functions will be run immediately. We may want
to `setImmediate` these instead for proper ordering? But I think it's
okay since the init functions for these particular subscriptions will
never run, and we don't assume ordering between separate subscriptions.