Remove advance_cursor_to_next_integer_point as egui no longer has the issue it was needed for
The function was needed when egui used to leave cursor at .5 float offset,
which was bad as smallest possible float inaccuracy downwards would cause
it to round down instead of up, as egui rounds text positions to exact
pixel offset as that looks better in the end.

Now that egui leaves cursor at .0 during normal usage, now this function
became problematic as smallest inaccuracy upwards would cause it to ceil()
to next integer, causing flickering when there was none. (And the function
didn't actually move cursor to the next point but also implicitly added
several points of spacing, which made things even worse)