Focus the replay entry in the file browser when Watch Replay is clicked. (#1116)
This is mostly useful if you click on Watch Replay button without
focusing that replay entry first. The replay would start fine, but the
file entry would not be focused, confusing you what was the last replay
you watched.

I couldn't just remove the `event.stopPropagation()` on the Watch Replay
button, because that would also expand the replay info panel as well as
started the replay.