Remove party feature.
Currently this feature is somewhat degenerate with player behavior and
is hard to support with a small playerbase, and is not fully realized
anyway (we can't use parties to form lobbies/watch replays/etc.). Since
we don't plan to support parties in many of the team-based matchmaking
modes, it also makes things very confusing in that regard (and adds a
ton of complexity to the matchmaker).

This isn't necessarily a commit that can be cleanly rolled back, as
near-future changes to other related code won't take it into account.
If/when we do want this in the future, this code may be a good starting
point, or not, unsure.