2025 Design Refresh (#1141)
Rework most of the app to fit within a new design that will enable us to expand our featureset more easily going forward. With this new design, we no longer place "activities" into a sidebar. Instead, they are top-level full-window pages (which should make designing UIs for them that work well much easier, and also makes it possible to directly link to them).

This redesign was necessary to support more matchmaking types going forward, as well as multi-type queuing (neither of these are done as part of this work, but will be done in the future). As part of this work I tried to take a "default to available" approach to signed out users, so many more things are available now (and will also be visible to search engine crawlers). Some pages might need to be restructured to be more amenable to crawlers, but we can do that down the line. Many more things are also available while offline (too many things, this will need work in follow-up commits).

There is a bunch of work that is un-done as of these commits. I've filed some bugs for some of them, but also have a todo list of things I plan to work through in the very near future. Once we get through most of those we'll probably need to run through all the different app flows to double-check that things are functioning correctly and look halfway decent.