Migrate notification popups to motion.
I removed the scrollability of the container here because it seemed kind
of annoying to maintain (I needed to size up the container as it was
cutting off the exiting notification, and thus needed to change the flex
properties on it in a way that made it not overflow). It didn't seem
particularly useful (you'd rarely have that many notifications visible
since they dismiss after 10 seconds anyway) and could view more by
dismissing specific ones.