Add react-compiler.
We had a number of issues that it linted, mostly around misusing refs in
render. Some of these feel fairly hard to fix without a deeper re-think,
so I've silenced a number of the warnings for now. Generally anywhere
we're using usePrevious is a problem, but there is probably some other
approach that would work if we restructure things a bit.

This stuff doesn't need to be immediately cleaned up, the compiler will
just not optimize those components, but long-term we should probably try
to remove usages of these problematic hooks and remove them altogether.

In general at this point you shouldn't need to
useCallback/useStableCallback any more, as the compiler will memoize for
you. Mostly try to put function declarations inline with the component
if they are short enough/only used once, as it will avoid needing to
declare parameter types explicitly and keep the implementation near the
usage.

The react-hooks exhaustive-deps lint rule will fire erroneously if you
use a non-`useCallback`'d function as a `useEffect` dep. Unfortunately
this lint rule does 2 things: ensure you memoize functions (which
doesn't have any awareness of react-compiler), and ensure that all your
`useEffect` dependencies are actually named. We want the latter to still
happen (I've tested and useEffect will indeed not fire enough without
naming all the deps explicitly), so we can't turn off the rule entirely.
Hopefully they will fix this in the future, for now I would just
`useCallback` those specific functions, as it's sort of rare we do this
anyway.