Create a news feed for the homepage (static content for now).
I've added a few images that will be automatically chosen based on the
index of the news entry. The expectation is that any new entries we add
(probably just one for each new version) will be added at the end of the
array so indexes will remain static for the older entries. At some
point in the future we'll finish the dynamic news feed work and that can
replace this content.

I added bottom links to the homepage (and associated pages) as well, so
that we have a place to link the policies, socials, and FAQ. FAQ has
been updated to fit the new design better.

Since the changelog is available as news entries, we no longer need to
maintain/use the markdown version, so it has been deleted (along with
the changelog dialog and associated code).