Ban all connected identifiers for a user in one DB query.
Previously we would queue up separate queries for each connected user
(along with dealing with the session stuff + websockets separately),
which, for users with ~100 accounts, means ~100 separate DB transactions
running concurrently. This pretty quickly overloads the DB pool, which
means a lot of queries are left waiting and we get a bunch of logs about
slow queries.

The new setup rearranges some things in a way that potentially allows
for a race, but is a much simpler/faster process that only needs a
single DB transaction. In practice I don't think the potential race
really has big consequences anyway (it basically self-corrects should
that user continue using the app).