Implement a restricted names feature.
These are added via an admin interface, can currently either be an exact
match (one specific name) or regex. These are checked via server-rs,
with the TS server talking to it over a local HTTP request.

Currently this doesn't affect users that have names that match
restrictions (we haven't built the name change feature yet), but we can
write something to fix that up in the future when it exists.