Move whisper init to just give user IDs instead of trying to return SbUsers the model method created itself, which will make it easier to add fields to SbUser in the future.
The name would be out of date anyway if that user changed their name
while this user was connected + connected a new client. I could have
fixed this by making the websocket subscription query for an up-to-date
SbUser, but it seemed somewhat wasteful/useless since the UI components
we use to display these are able to batch request this info anyway.