Only allow signup in the app.
This will be better for smurf detection + ban evasion, so it seems
preferred. There is still more work that needs to be done (as far as
enforcing identifiers more, removing web-specific code in signup
actions, making email verification work without needing to click links,
etc.). I'm about to redo the various auth pages for the new design, so
I'll do those pieces then.

This commit adjusts the the integration tests to override headers to make
them able to sign up (despite being in the web version), and provides the
tests with a route that will show them the "real" signup page (if users
in prod happen to see this it's fine, the server will still prevent them
from proceeding.