Implement animations for play button.
This moves the play button from an SVG into (mostly) HTML/CSS since it
makes it easier to animate. It also turns the gradient into a simpler
gradient and then uses 2 blurred circles in the corners to make a
gradient that can animate.

2 animations are currently present:
- A "breathing" animation that happens when the user is not interacting
  with the button, optionally faster for things like being in the
  matchmaking queue
- An animation that subtly tracks the mouse cursor when the user is
  hovering the button

I think it might be nice to have a click animation utilizing these same
pieces but I haven't tried to make that yet.

Ideally we'd also add click sounds for this button (and probably all the
appbar buttons), but I also haven't pursued that yet.