Implement urgent messages.
Urgent messages are settable by any admins with `manageNews`
permissions, and will notify any connected clients over websockets that
a new message has been set. In the near future we can use that
notification to show a pip on the home button, but for now they'll just
see it when they refresh.

Clients retrieve the urgen message via graphql (since the other home
content will also be over graphql).

The message has a title (showed on the home page) and markdown contents
(shown in a dialog, I created MarkdownDialog for doing this type of
thing).

This commit also restructures the admin dashboard a bit so it's easy to
maintain. I also fixed usages of TypedEmitter, as how we were using it
previously was no longer erroring with incorrect event names (I think
this was a TS change, dunno).